perf(graphic): hide only sprites freed since the previous frame
NGGraphicSystemDraw unconditionally swept every unused entity and UI
sprite each frame (~250-320 SCB3 writes inside the VBlank window, even
with nothing changing). Track the previous frame's allocation high-water
marks and clear only the freed delta; sprites below the marks are
rewritten by their new owners during the draw pass.

NGSpriteHideRange takes a u16 count now, dropping the 255-sprite
batching loops, and NGGraphicSystemReset hides the sprite list in one
call.

+++ hal/include/ng_sprite.h
@@ -153,9 +153,9 @@ static inline u16 NGSpriteSCB4(s16 screen_x) {
  * Uses optimized VRAM macros for fast clearing.
  *
  * @param first_sprite First sprite index to hide
- * @param count Number of sprites to hide
+ * @param count Number of sprites to hide (may span the whole sprite list)
  */
-static inline void NGSpriteHideRange(u16 first_sprite, u8 count) {
+static inline void NGSpriteHideRange(u16 first_sprite, u16 count) {
     if (count == 0)
         return;
     NG_VRAM_DECLARE_BASE();
